% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemrest_core.R
\name{check_and_install}
\alias{check_and_install}
\title{Check and Install Required Packages}
\usage{
check_and_install(pkg)
}
\arguments{
\item{pkg}{A character string naming the package to load/install.}
}
\value{
Loads the requested package into the session.
}
\description{
Installs and loads a package if not already installed.
}
\keyword{internal}
