% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efa_diagnostics.R
\name{determine_n_factors}
\alias{determine_n_factors}
\title{Determine the number of factors using Parallel Analysis.}
\usage{
determine_n_factors(
  data,
  cor_method = "pearson",
  missing = "listwise",
  pd_action = "fail",
  parallel_iterations = 100L,
  correlation_cache = NULL,
  parallel_method = "fa"
)
}
\description{
Determine the number of factors using Parallel Analysis.
}
\keyword{internal}
