% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efa_diagnostics.R
\name{efa_custom}
\alias{efa_custom}
\title{Run an EFA with correlation and numerical diagnostics.}
\usage{
efa_custom(
  data,
  n_factors = 1,
  cor_method = "polychoric",
  extract = "uls",
  rotate = "oblimin",
  missing = "listwise",
  pd_action = "fail",
  reliability = "alpha_omega",
  correlation_cache = NULL
)
}
\description{
Run an EFA with correlation and numerical diagnostics.
}
\keyword{internal}
