% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemrest_core.R
\name{howard}
\alias{howard}
\title{Check for cross-loading using Howard and Costello & Osborne criteria.}
\usage{
howard(primary, secondary)
}
\description{
Check for cross-loading using Howard and Costello & Osborne criteria.
}
\keyword{internal}
