% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemrest_core.R
\name{identify_problem_items}
\alias{identify_problem_items}
\title{Identify low-loading and cross-loading items.}
\usage{
identify_problem_items(efa_res, min_loading = 0.3, loading_diff = 0.1)
}
\description{
Identify low-loading and cross-loading items.
}
\keyword{internal}
