% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemrest_core.R
\name{cor_matrix_custom}
\alias{cor_matrix_custom}
\title{Calculate a correlation matrix.}
\usage{
cor_matrix_custom(data, method = "polychoric")
}
\description{
Calculate a correlation matrix.
}
\keyword{internal}
