% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemrest_core.R
\name{descriptive_stats}
\alias{descriptive_stats}
\title{Calculate basic descriptive statistics for a dataset.}
\usage{
descriptive_stats(data)
}
\description{
Calculate basic descriptive statistics for a dataset.
}
\keyword{internal}
