% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemrest_core.R
\name{determine_n_factors}
\alias{determine_n_factors}
\title{Determine the number of factors using Parallel Analysis.}
\usage{
determine_n_factors(data, cor_method = "pearson")
}
\description{
Determine the number of factors using Parallel Analysis.
}
\keyword{internal}
