% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemrest_core.R
\name{efa_custom}
\alias{efa_custom}
\title{Run a custom EFA.}
\usage{
efa_custom(
  data,
  n_factors = 1,
  cor_method = "polychoric",
  extract = "uls",
  rotate = "oblimin"
)
}
\description{
Run a custom EFA.
}
\keyword{internal}
