% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemrest_core.R
\name{get_combinations}
\alias{get_combinations}
\title{Generate combinations of items to remove.}
\usage{
get_combinations(items)
}
\description{
Generate combinations of items to remove.
}
\keyword{internal}
