% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemrest_core.R
\name{identify_problem_items}
\alias{identify_problem_items}
\title{Identify problematic items (low-loading or cross-loading).}
\usage{
identify_problem_items(
  efa_res,
  primary_cutoff = 0.4,
  secondary_max = 0.3,
  diff_min = 0.2,
  low_loading_thresh = 0.3
)
}
\description{
Identify problematic items (low-loading or cross-loading).
}
\keyword{internal}
