% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemrest_core.R
\name{itemrest}
\alias{itemrest}
\title{Evaluate Item Removal Strategies for Exploratory Factor Analysis (EFA)}
\usage{
itemrest(
  data,
  cor_method = "pearson",
  n_factors = NULL,
  extract = "uls",
  rotate = "oblimin"
)
}
\arguments{
\item{data}{A numeric \code{data.frame} or \code{matrix} for the analysis.}

\item{cor_method}{The correlation method to use, e.g., \code{"pearson"} or \code{"polychoric"}.}

\item{n_factors}{The number of factors. If \code{NULL}, it's determined automatically by parallel analysis.}

\item{extract}{The factor extraction (estimation) method. See \code{psych::fa}. Default is \code{"uls"}.}

\item{rotate}{The rotation method. See \code{psych::fa}. Default is \code{"oblimin"}.}
}
\value{
An object of class \code{itemrest_result}.
}
\description{
This function identifies low-quality items (low-loading or cross-loading)
based on initial EFA results, then tests different combinations of removing
these items to find optimal model fit. It returns an object containing all
results. It only prints progress information to the console during computation.
}
