% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemrest_core.R
\name{print.itemrest_result}
\alias{print.itemrest_result}
\title{Print method for \code{itemrest_result} class}
\usage{
\method{print}{itemrest_result}(x, report = "optimal", ...)
}
\arguments{
\item{x}{An object of class \code{itemrest_result}.}

\item{report}{The type of report to generate: \code{"optimal"} (default) or \code{"all"}.}

\item{...}{Other arguments (not used).}
}
\description{
Print method for \code{itemrest_result} class
}
