% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemrest_core.R
\name{test_removals}
\alias{test_removals}
\title{Test removal strategies and build a summary table.}
\usage{
test_removals(
  data,
  base_items,
  combs,
  n_factors,
  cor_method,
  extract,
  rotate,
  min_loading,
  loading_diff
)
}
\description{
Test removal strategies and build a summary table.
}
\keyword{internal}
